dockers/Florencev2/pyproject.toml
[project]
name = "florencev2"
version = "0.1.0"
description = ""
authors = [
    {name = "Victor Manuel Martinez Garcia",email = "victor.martinez@itcl.es"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pillow (>=11.2.1,<12.0.0)",
    "timm (>=1.0.15,<2.0.0)",
    "einops (>=0.8.1,<0.9.0)",
    "transformers (>=4.51.3,<5.0.0)",
    "safetensors (>=0.5.3,<0.6.0)",
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.1,<0.35.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
